+++ seantis/people/supermodel/__init__.py
@@ -10,7 +10,11 @@ from .schemahandler import (
     get_detail_fields,
     set_detail_fields,
     get_custom_column_titles,
-    set_custom_column_titles
+    set_custom_column_titles,
+    get_list_render_options,
+    set_list_render_options,
+    get_detail_render_options,
+    set_detail_render_options
 )
 
 from .schema_columns import (